Unbreak mips64 build by MFCing r202031 that fixes _ALIGN macros
Former-commit-id: be5734f61a7983992039541d02cfffb87f261c0f

+++ sys/mips/include/param.h
@@ -169,7 +169,7 @@ directive|endif
 end_endif
 
 begin_comment
-comment|/*  * Round p (pointer or byte index) up to a correctly-aligned value for all  * data types (int, long, ...).	  The result is u_int and must be cast to  * any desired pointer type.  */
+comment|/*  * Round p (pointer or byte index) up to a correctly-aligned value for all  * data types (int, long, ...).	  The result is u_long and must be cast to  * any desired pointer type.  */
 end_comment
 
 begin_define
@@ -186,7 +186,7 @@ name|_ALIGN
 parameter_list|(
 name|p
 parameter_list|)
-value|(((u_int)(p) + _ALIGNBYTES)&~ _ALIGNBYTES)
+value|(((u_long)(p) + _ALIGNBYTES)&~ _ALIGNBYTES)
 end_define
 
 begin_define
